Reorganized horse riding AI
